Move logic into App
